refactor(*): controller, util 패키지 생성 및 각 의미에 맞게 클래스 이동
